refactor: rename cubit
