+++ modules/features/onboarding/lib/src/presentation/extension/extension.dart
@@ -1,0 +1,1 @@
+export './api/api_result.dart';
