+++ modules/features/onboarding/lib/src/presentation/view/home_view.dart
@@ -2,7 +2,7 @@ import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
-import 'package:onboarding/src/presentation/bloc/list_photo_cubit.dart';
+import '../../presentation/cubit/list_photo_cubit.dart';
 
 class HomeView extends StatefulWidget {
   const HomeView({super.key});
